test: move Redis storage fake to mocks
