+++ apps/web/modules/auth/lib/better-auth-session-fallback.test.ts
@@ -1,45 +1,12 @@
-import { type BetterAuthOptions, betterAuth } from "better-auth";
+import { REDIS_ERROR, createSecondaryStorageMock } from "./__mocks__/secondary-storage.mock";
+import { betterAuth } from "better-auth";
 import { memoryAdapter } from "better-auth/adapters/memory";
 import { describe, expect, test, vi } from "vitest";
 
 const BASE_URL = "https://app.formbricks.test";
-const REDIS_ERROR = new Error("Socket closed unexpectedly");
-
-const createSecondaryStorage = () => {
-  const values = new Map<string, string>();
-  let failReads = false;
-
-  return {
-    storage: {
-      get: vi.fn(async (key: string) => {
-        if (failReads) throw REDIS_ERROR;
-        return values.get(key) ?? null;
-      }),
-      getAndDelete: vi.fn(async (key: string) => {
-        const value = values.get(key) ?? null;
-        values.delete(key);
-        return value;
-      }),
-      increment: vi.fn(async (key: string) => {
-        const value = Number(values.get(key) ?? 0) + 1;
-        values.set(key, String(value));
-        return value;
-      }),
-      set: vi.fn(async (key: string, value: string) => {
-        values.set(key, value);
-      }),
-      delete: vi.fn(async (key: string) => {
-        values.delete(key);
-      }),
-    } satisfies BetterAuthOptions["secondaryStorage"],
-    failReads: () => {
-      failReads = true;
-    },
-  };
-};
 
 const createAuthInstance = (storeSessionInDatabase: boolean) => {
-  const secondaryStorage = createSecondaryStorage();
+  const secondaryStorage = createSecondaryStorageMock();
   const log = vi.fn();
   const auth = betterAuth({
     baseURL: BASE_URL,
